refactor: move sourceset code out of tech (#5037)
